// src/canisters/yuku-old/yuku_origyn_art/module_6a35faf5/origyn_art_6a34faf5.did.d.ts
export interface CollectionMetadata {
    artworkMedium: string;
    artworkDescription: string;
    media: Array<string>;
    artworkList: string;
    artworkSize: string;
    authorPicture: string;
    artworkDetails: string;
    name: string;
    description: string;
    authorBirth: string;
    coverImage: string;
    copyright: string;
    artAuthor: string;
    authorProvenance: string;
}
// ...
export interface ICTokenSpec {
    fee: bigint;
    decimals: bigint;
    canister: Principal;
    standard: { ICRC1: null } | { EXTFungible: null } | { DIP20: null } | { Ledger: null };
    symbol: string;
}
export interface ICTokenSpec__1 {
    fee: bigint;
    decimals: bigint;
    canister: Principal;
    standard: { ICRC1: null } | { EXTFungible: null } | { DIP20: null } | { Ledger: null };
    symbol: string;
}
// ...
export interface PrimarySale {
    fee: { rate: bigint; precision: bigint };
    egg_canister: Principal;
    token: ICTokenSpec__1;
    owner: Principal;
    end_time: Time;
    available: bigint;
    threshold_trade: bigint;
    start_time: Time;
    threshold_percentage: bigint;
    fraction_canister: Principal;
    unit_price: bigint;
    volumeTrade: bigint;
    total_supply: bigint;
}
// ...
export type Time = bigint;
// export type TokenIdentifier = string;
// export type User = { principal: Principal } | { address: AccountIdentifier__1 };

export default interface _SERVICE {
    // addCollectionMetadata: (arg_0: Principal, arg_1: CollectionMetadata) => Promise<undefined>;
    // addManager: (arg_0: Array<Principal>) => Promise<undefined>;
    // addMintedNFT: (arg_0: Principal, arg_1: Array<string>) => Promise<undefined>;
    // addOgyCanisters: (arg_0: Array<Principal>) => Promise<undefined>;
    // addPointSettlementV: (arg_0: Array<[Principal, bigint]>) => Promise<undefined>;
    // addPrimarySale: (arg_0: PrimarySale) => Promise<Result_2>;
    // addPrimarySaleBlock: (arg_0: Principal, arg_1: boolean) => Promise<undefined>;
    // addSupportedTokens: (arg_0: Array<ICTokenSpec>) => Promise<undefined>;
    // addWhitelist: (arg_0: Array<Principal>) => Promise<undefined>;
    // batchSettleRecord: (arg_0: Array<bigint>) => Promise<undefined>;
    // buy_primary_nft_origyn: (arg_0: Principal, arg_1: bigint) => Promise<Array<Result_1>>;
    // delManager: (arg_0: Array<Principal>) => Promise<undefined>;
    // delOgyCanisters: (arg_0: Array<Principal>) => Promise<Array<Principal>>;
    // delPrimarySale: (arg_0: Principal) => Promise<undefined>;
    // delwhitelist: (arg_0: Array<Principal>) => Promise<undefined>;
    // flushPointSettlement: () => Promise<undefined>;
    // flushRecordSettlement: () => Promise<undefined>;
    // getCanisterLog: (arg_0: [] | [CanisterLogRequest]) => Promise<[] | [CanisterLogResponse]>;
    // getManager: () => Promise<Array<Principal>>;
    // getOwner: () => Promise<Principal>;
    // getPointSettlements: () => Promise<Array<[bigint, PointSale]>>;
    // getRecordSettlement: () => Promise<Array<[bigint, RecordSettle]>>;
    // getStagedTokens: (arg_0: Principal) => Promise<Array<string>>;
    // getrecordMarks: () => Promise<Array<[Principal, Array<[string, bigint]>]>>;
    // getwhitelist: () => Promise<Array<Principal>>;
    // handleSecondPoint: (arg_0: Principal) => Promise<undefined>;
    listCollections: () => Promise<Array<[Principal, CollectionMetadata]>>;
    // listMintedAndStaged: (
    //     arg_0: Principal,
    // ) => Promise<{ staged: [] | [Array<string>]; minted: [] | [Array<string>] }>;
    listOgyCanisters: () => Promise<Array<Principal>>;
    listPrimarySale: () => Promise<Array<PrimarySale>>;
    listSupportedTokens: () => Promise<Array<ICTokenSpec>>;
    // mintFractionNFTByEggCanister: (
    //     arg_0: Principal,
    //     arg_1: { rate: bigint; precision: bigint },
    // ) => Promise<Array<Result>>;
    // rmSupportedTokens: (arg_0: Array<string>) => Promise<Array<ICTokenSpec>>;
    // setOwner: (arg_0: Principal) => Promise<undefined>;
    // setPointSettlements: (arg_0: bigint, arg_1: Principal, arg_2: bigint) => Promise<undefined>;

    // settlePoint: (arg_0: bigint) => Promise<SettlePointResult>;
    // settleRecord: (arg_0: bigint) => Promise<SettleRecordResult>;
}